Fix up of memory leak

+++ cpp/foxglove/tests/test_server.cpp
@@ -259,7 +259,7 @@ TEST_CASE("Subscribe and unsubscribe callbacks") {
     subscribe_calls.push_back(channel_id);
     cv.notify_all();
   };
-  callbacks.onUnsubscribe = [&](uint64_t channel_id, uint64_t client_id [[maybe_unused]]) {
+  callbacks.onUnsubscribe = [&](uint64_t channel_id, uint32_t client_id [[maybe_unused]]) {
     std::scoped_lock lock{mutex};
     unsubscribe_calls.push_back(channel_id);
     cv.notify_all();
